requirements.txt
confluent-kafka==2.3.0
py4j==0.10.9.7
pyspark==3.5.0
simplejson==3.19.2
boto3==1.34.59
python-dotenv==1.0.1